refactor: 提取StoreCommand基于git进行命令
